+++ ai/Algorithm/PathfinderRRT.py
@@ -20,7 +20,7 @@ from RULEngine.Util.Position import Position
 from RULEngine.Util.constant import POSITION_DEADZONE
 from ai.Algorithm.IntelligentModule import Pathfinder
 
-from ai.Debug.DebugInterface import COLOR_ID_MAP, DEFAULT_PATH_TIMEOUT
+from ai.Debug.debug_interface import COLOR_ID_MAP, DEFAULT_PATH_TIMEOUT
 
 OBSTACLE_DEAD_ZONE = 700
 TIME_TO_UPDATE = 1
