+++ ai/Debug/debug_command.py
