+++ ai/Debug/ui_debug_command.py
