+++ ai/STA/Tactic/GoToPosition.py
@@ -4,12 +4,12 @@ from ai.STA.Tactic.Tactic import Tactic
 from ai.STA.Tactic.tactic_constants import Flags, DEFAULT_TIME_TO_LIVE
 from ai.STA.Action.MoveTo import MoveTo
 from ai.STA.Action.Idle import Idle
-from ai.states.ModuleState import NonExistentModule
+from ai.states.module_state import NonExistentModule
 from ai.Util.geometry import get_distance, get_angle
 from RULEngine.Util.Pose import Pose
 from RULEngine.Util.constant import ANGLE_TO_HALT, POSITION_DEADZONE, PLAYER_PER_TEAM
 
-from ai.states.ModuleState import ModuleState
+from ai.states.module_state import ModuleState
 
 __author__ = 'RoboCupULaval'
