+++ ai/executors/play_executor.py
@@ -1,6 +1,6 @@
 # Under MIT License, see LICENSE.txt
 
-from ai.executors.Executor import Executor
+from ai.executors.executor import Executor
 
 class PlayExecutor(Executor):
